### 2022-01-13 #4063 @whedon
[PRE REVIEW]: PyPVRPM: Photovoltaic Reliability and Performance Model in Python
**Submitting author:** @brandons209 (<a href="http://orcid.org/0000-0001-8278-7158">Brandon Silva</a>)
**Repository:** <a href="https://github.com/FSEC-Photovoltaics/pvrpm-lcoe" target ="_blank">https://github.com/FSEC-Photovoltaics/pvrpm-lcoe</a>
**Version:** v1.6.0
