### 2022-01-13 #4063 @danielskatz
👋 @brandons209 - I've suggested a bunch of changes in  https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/pull/26 to fix the bib errors
