### 2022-01-14 #4063 @brandons209
> wave @brandons209 - I've suggested a bunch of changes in [FSEC-Photovoltaics/pvrpm-lcoe#26](https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/pull/26) to fix the bib errors

Thank you! The paper compiled fine so I didn't realize there were errors. I'll merge the changes.
