### 2022-01-14 5aab014
fix doi and remove duplicate citation
