+++ paper.bib
@@ -14,14 +14,6 @@
   url = {https://github.com/nrel/pysam}
 }
 
-@article{PVRPM,
-  url = {https://www.nrel.gov/docs/fy19osti/72212.pdf},
-  Author = {Janine Freeman and Geoffrey Klise and Andy Walker and Olga Lavrova},
-  Journal = {WCPEC-7},
-  Title = {Evaluating Energy Impacts and Costs from {PV} Component Failures},
-  Year = 2018
-}
-
 @article{PVRPM:2011,
   url = {https://energy.sandia.gov/wp-content/gallery/uploads/PV-Rollout_PVRPM_SAND2011-9293C.pdf},
   Author = {Jennifer Granata and Steven Miller and Joshua Stein},
@@ -35,6 +27,7 @@
   Author = {Steven Miller and Jennifer Granata and Joshua Stein},
   number = {SAND2012-10342},
   Title = {The Comparison of Three Photovoltaic System Designs Using the Photovoltaic Reliability and Performance Model ({PVRPM})},
+  doi = {10.2172/1088066},
   Year = 2012
 }
 
