+++ paper.md
@@ -34,7 +34,7 @@ The ability to perform accurate techno-economic analysis of photovoltaic (PV) sy
 
 # Statement of need
 
-As photovoltaic technology becomes cheaper and more widely used, solutions are needed to accurately assess their performance and monitor them to safeguard performance and reduce downtime. Existing PV performance models assume ideal operation without considering any failures or repairs. In contrast, the value of monitoring at different levels (e.g., inverter, combiner, string, or module-level monitoring) is unclear since these depend on the quality of components, financial metrics, and climatic conditions. PVRPM was initially developed in Goldsim [@PVRPM:2011] [@PVRPM:2012] and then adapted in LK script [@osti_1761998] to improve the accuracy of energy yield simulations by inclusion of realistic energy losses and O&M costs [@PVRPM]. However, these models lack many needed features such as efficient matrix operations, command-line interfaces, and external libraries to perform computations.
+As photovoltaic technology becomes cheaper and more widely used, solutions are needed to accurately assess their performance and monitor them to safeguard performance and reduce downtime. Existing PV performance models assume ideal operation without considering any failures or repairs. In contrast, the value of monitoring at different levels (e.g., inverter, combiner, string, or module-level monitoring) is unclear since these depend on the quality of components, financial metrics, and climatic conditions. PVRPM was initially developed in Goldsim [@PVRPM:2011] [@PVRPM:2012] and then adapted in LK script [@osti_1761998] to improve the accuracy of energy yield simulations by inclusion of realistic energy losses and O&M costs [@PVRPM:EVAL]. However, these models lack many needed features such as efficient matrix operations, command-line interfaces, and external libraries to perform computations.
 
 Furthermore, the LK-based PVRPM does not include simulation of monitoring configurations (e.g., inverter-level vs. string-level monitoring), which PyPVRPM implements. PyPVRPM takes the original logic from PVRPM, refines it, and adds new features to make it faster, easily configurable, more powerful. This is done using established Python libraries such as NumPy and Pandas. PyPVRPM can also batch simulate in an automated manner on a computer cluster. PyPVRPM can simulate any component of a PV plant from the module level up to the grid while also considering different monitoring scenarios.
 
