### 2022-01-22 #4093 @whedon
[REVIEW]: PyPVRPM: Photovoltaic Reliability and Performance Model in Python
**Submitting author:** @brandons209 (<a href="http://orcid.org/0000-0001-8278-7158">Brandon Silva</a>)
**Repository:** https://github.com/FSEC-Photovoltaics/pvrpm-lcoe
**Branch with paper.md** (empty if default branch):
**Version:** 1.7.4

- [x] #4213
