### 2022-01-22 #4063 @brandons209
> Hi @brandons209 can suggest a few names how may be able to review this? My search so far have yielded a pretty small and interconnected group of people.

Hello! I am not sure what you are asking. Are you asking me to recommend some people outside of JOSS to help review this? I can recommend some people from NREL if that's the case.
