### 2022-01-24 #4093 @jgostick
Hi @blthayer and @SarthakJariwala.  Thank you for being a reviewer for JOSS.  At the top of this thread you will see a checklist for you to use as a guideline for your review.  At JOSS we generally work with the authors to help them get their submission into publishable form (if it is not already), so you can interact with them here or on their repo regarding any changes you think are needed.  Remember that the review is publish and non-anonymous, so is meant to be a collaborative effort, and your github handles will appear on the final PDF as reviewers, so you'll get credit for your efforts.

Let's aim to have this done in a about 1 month?  So Feb 28th?
