+++ setup.py
@@ -3,10 +3,10 @@ from setuptools import setup, find_packages
 setup(
     name="pvrpm",
     description="Run simulations to estimate LCOE costs for PV systems using SAM.",
-    version="1.6.0",
+    version="1.7.0",
     packages=find_packages(include=["pvrpm", "pvrpm.*"]),
     install_requires=[
-        "nrel-pysam==2.2.3",
+        "nrel-pysam==3.0.0",
         "pandas",
         "click",
         "pyyaml",
