### 2022-02-05 #4093 @blthayer
@jgostick - it looks like I need a fresh invitation sent in order to be able to edit my checklist. When I went to the following URL, it said the invitation was expired: https://github.com/openjournals/joss-reviews/invitations

This isn't (yet) blocking me from performing my review.

Thanks!
