### 2022-02-05 FSEC-Photovoltaics/pvrpm-lcoe#27 @blthayer
DOC - Complete "SAM/PVRPM Assumptions / Limitations" documentation
Title says it all :smile:

For your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/4093), it'd be nice to have these assumptions/limitations [documented](https://pvrpm.readthedocs.io/en/latest/tutorial_assumptions_limitations.html). A high-level first pass would be adequate for me.
