### 2022-02-05 FSEC-Photovoltaics/pvrpm-lcoe#28 @blthayer
DOC - "Getting Started" section should link to SAM
For users who don't know what SAM is nor have experience with it, it would be helpful if the [Getting Started](https://pvrpm.readthedocs.io/en/latest/tutorial_getting_started.html) page gave a link to [SAM](https://sam.nrel.gov/).
