### 2022-02-05 FSEC-Photovoltaics/pvrpm-lcoe#29 @blthayer
DOC - "Installation" before "Getting Started?"
Currently, the [documentation](https://pvrpm.readthedocs.io/en/latest/index.html) is organize roughly like:

- Assumptions
- Getting Started
- Installation
- Logic Diagram

However, the "Getting Started" section seems to infer that installation has already occurred, as it references both SAM procedures and PVRPM configuration etc. I propose moving the "Installation" section to _before_ the "Getting Started" section
