### 2022-02-05 FSEC-Photovoltaics/pvrpm-lcoe#30 @blthayer
DOC - SAM version
The [GitHub README](https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/tree/c6cd62a8de6576ff16e2d7355534f9ab216ac7eb) indicates SAM version `2021.12.02` while the [documentation on Read the Docs](https://pvrpm.readthedocs.io/en/latest/tutorial_installation.html) specifies version `2020.11.29`.
