### 2022-02-05 FSEC-Photovoltaics/pvrpm-lcoe#31 @blthayer
DOC - Module Order
The [Module Order](https://pvrpm.readthedocs.io/en/latest/tutorial_getting_started.html#module-order) documentation seems to infer that I should look up my module order in a gigantic table from [here](https://nrel-pysam.readthedocs.io/en/master/Configs.html#sam-simulation-configurations).

However, it's much simpler to do something like the following:

```python
>>> import json
>>> with open("my_sam_case.json", "r") as f:
...     data = json.load(f)
>>> modules = {x: data[x] for x in data.keys() if x.startswith("compute_module")}
>>> modules
{'compute_module_0': 'pvwattsv8', 'compute_module_1': 'belpe', 'compute_module_2': 'grid', 'compute_module_3': 'utilityrate5', 'compute_module_4': 'cashloan'}
```
