### 2022-02-05 FSEC-Photovoltaics/pvrpm-lcoe#32 @blthayer
DOC - Link to sample YAML up front
The [Configuring PVRPM](https://pvrpm.readthedocs.io/en/latest/tutorial_getting_started.html#configuring-pvrpm) section of the documentation sort of just jumps in to describing sections of the PVRPM configuration file.

As a user, I'd like to be given/linked to an example configuration file up front so that I can easily follow along with the documentation and update (without writing from scratch) the relevant sections as I go.

It looks like there's already an [example file](https://pvrpm.readthedocs.io/en/latest/example_pvrpm_config.html), so this issue could be addressed by:
- Link to file early in configuration section of documentation
- Provide a raw download link for the example file. Currently, you have to manually highlight the gigantic box, copy + paste into a new file. Direct download would be faster and easier.
