### 2022-02-05 FSEC-Photovoltaics/pvrpm-lcoe#33 @blthayer
DOC - General formatting suggestion
(100% optional) - In the documentation, italics is often used, *e.g.*:

> ... then open your *.sam* case file.

and

> *can_fail* can be set to true or false...

It's easier to read if fixed width formatting (single back-ticks in Markdown, double back-ticks in ReStructuredText) is used, *e.g.*:
- ... then open you `.sam` case file
- `can_fail` can be set to `true` or `false`
