### 2022-02-05 FSEC-Photovoltaics/pvrpm-lcoe#34 @blthayer
DOC - typo
Related to the Weibull distribution in the [Distributions for Failures, Repairs, Monitoring](https://pvrpm.readthedocs.io/en/latest/tutorial_getting_started.html#distributions-for-failures-repairs-monitoring) section:

> On the [Wikipedia page](https://en.wikipedia.org/wiki/Weibull_distribution) for the weibull distribution is the parameter *k lambda* is calculated from the mean.

Are both parameters calculated from the mean?
