### 2022-02-05 FSEC-Photovoltaics/pvrpm-lcoe#35 @blthayer
DOC - decay nitpick
In the [Failure Setup](https://pvrpm.readthedocs.io/en/latest/tutorial_getting_started.html#failure-setup) portion of the documentation, it states:

> If you set *decay_fraction* to 0.5, then at first, 50 percent of the components will fail with this failure mode, then 25 percent, then 12.5 percent, etc., until it goes to 0.

As I said in the title, this is a nitpicky suggestion :smile: : This won't actually ever go to zero, but rather approaches zero.
