### 2022-02-05 FSEC-Photovoltaics/pvrpm-lcoe#36 @blthayer
BUG (maybe) - Error loading and verifying case
# Overview
When attempting to run `pvrpm`, I get the following error:
```bash
$ pvrpm run --case ./ ./richland.yaml
2022-02-05 11:22:12,974--ERROR: Error loading and verifying case: Couldn't find module for file richland.json!
```

# System/Environment
- OS: Ubuntu 20.04.3
- Python: 3.10.2
- SAM: 2021.12.02
- pvrpm installed from master at c6cd62a8de6576ff16e2d7355534f9ab216ac7eb

# Reproducing

Unzip [richland.zip](https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/files/8008738/richland.zip) into a directory. It contains:
- richland.sam
- richland.json (exported data from SAM via "Generate Code" -> "JSON for inputs")
- richland.yaml (pvrpm configuration file)

From within the directory with the files above, execute `pvrpm run --case ./ ./richland.yaml`

# Misc Notes

The PVRPM docs seem to indicate that SAM should generate multiple JSON files, but this is not the case for me.

Also, when I use SAM to export to `json`, SAM crashes, but does successfully generate a single `json` file as well as the `.h` and `.so` files. Here's the output from SAM in my terminal:

```bash
...

(sam.bin:11945): Gtk-CRITICAL **: 11:17:53.945: IA__gtk_widget_set_size_request: assertion 'height >= -1' failed

(sam.bin:11945): Gtk-CRITICAL **: 11:17:53.951: IA__gtk_widget_set_size_request: assertion 'height >= -1' failed

(sam.bin:11945): Gtk-CRITICAL **: 11:17:53.956: IA__gtk_widget_set_size_request: assertion 'height >= -1' failed

(sam.bin:11945): Gtk-CRITICAL **: 11:17:53.957: IA__gtk_widget_set_size_request: assertion 'height >= -1' failed
free(): double free detected in tcache 2
Aborted (core dumped)

```
