### 2022-02-05 #4093 @blthayer
I'll use this comment to maintain a checklist of the issues which I've filed over in the author's repository.

- [ ] https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/issues/27
- [ ] https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/issues/28
- [ ] https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/issues/29
- [ ] https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/issues/30
- [ ] https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/issues/31
- [ ] https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/issues/32
- [ ] (OPTIONAL) https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/issues/33
- [ ] https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/issues/34
- [ ] (OPTIONAL) https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/issues/35
- [ ] https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/issues/36
