### 2022-02-05 #4093 @blthayer
My review is currently blocked by https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/issues/36, @brandons209
