### 2022-02-05 FSEC-Photovoltaics/pvrpm-lcoe#36 @brandons209
It looks like you are using a PVWatts model. PyPVRPM only supports Detailed PV Models. If you switch your model to that it should resolve this issue.
