### 2022-02-07 FSEC-Photovoltaics/pvrpm-lcoe#31 @brandons209
I'm not sure what you mean. Are you saying to have users create a JSON file for loading module order? Or including a json file that dictates all possible module orders and orders them properly?
