+++ docs/tutorial_getting_started.rst
@@ -3,6 +3,8 @@ Getting Started
 .. toctree::
   :hidden:
 
+**Make sure to follow :doc:`Installation <tutorial_1installation>` and install SAM before continuing here!**
+
 PVRPM enhances SAM's models for PV systems to obtain a more accurate LCOE than its base simulation. It also allows studying the effects of monitoring and repair techniques on PV systems throughout their lifetime.
 
 To get started, create a new SAM case in the SAM GUI. From there, you **must choose the Detailed Photovoltaic model** for PVRPM to work. This is required because PVRPM need's specific parameters that only exist in this model.
@@ -16,17 +18,19 @@ Exporting the SAM Case
 PVRPM works by reading the information in your SAM case and the PVRPM YAML configuration to properly run the simulation. For the first step, you need to export your SAM case to JSON files, which PVRPM uses to read in your SAM case.
 
 To do this:
-1. Open the SAM GUI, then open your `.sam` case file.
+1. Open the SAM GUI, then open your ``.sam`` case file.
 2. Once it is open, click the drop-down menu for the case, which is located next to the case name on the top bar.
-3. Click `Generate Code`, then `JSON for inputs`.
+3. Click ``Generate Code``, then ``JSON for inputs``.
 4. A file explorer window will open, allowing you to select a place to save the JSON files. Select an empty folder.
 
-Once this is done, the selected folder will contain a few JSON files. The names of these files will follow the convention of `case-name_module.json` where `case-name` is the name of the case, and `module` is the module that JSON represents. Pay attention to what modules you have; you'll need to know that for the next part. You can remove the `.h` and `.so` files.
+Once this is done, the selected folder will contain a few JSON files. The names of these files will follow the convention of ``case-name_module.json`` where ``case-name`` is the name of the case, and ``module`` is the module that JSON represents. Pay attention to what modules you have; you'll need to know that for the next part. You can remove the ``.h`` and ``.so`` files.
 
 Configuring PVRPM
 ------------------------
 
 This will go over every configuration option to set up the case study step by step. The example configuration is also heavily commented on to help with the parameters. Also, please study the logic diagram as it can help when setting up the configuration here. *Also note all of the values listed in examples are entirely arbitrary and do not represent a realistic case.*
+
+You can download the example configuration file :download:`here <../pvrpm/config/example_config.yml>` or view the example configuration file :doc:`here <example_pvrpm_config>`.
 
 Module Order
 ----------------
@@ -34,7 +38,7 @@ These modules correspond to the modules listed in the JSON files you obtained. T
 
 To make the module order easy to set up, go to this website: https://nrel-pysam.readthedocs.io/en/master/Configs.html#sam-simulation-configurations
 
-Then, find the modules you have under `SSC Compute Modules` column in the table; the SAM configuration column should be one of the `Detailed PV Model` rows. Once you find one containing all your modules, put them in the order as they appear in the `SSC Compute Modules` column.
+Then, find the modules you have under ``SSC Compute Modules`` column in the table; the SAM configuration column should be one of the ``Detailed PV Model`` rows. Once you find one containing all your modules, put them in the order as they appear in the ``SSC Compute Modules`` column.
 
 .. code-block:: yaml
   :linenos:
@@ -69,7 +73,7 @@ The results folder and number of realizations can be overridden from the command
 
 Case Setup
 ----------------
-Set up the basics of the case. Set the number of trackers, combiners, and transformers. Set `num_trackers` to `0` if you are not using trackers. The worst-case tracker can be set to true if you are using trackers. This means that failures in tracking components result in them being stuck in the worst way: the module is pointing the opposite of the sun's travel arc.
[…]
