### 2022-02-12 FSEC-Photovoltaics/pvrpm-lcoe#36 @blthayer
Oops, I missed that note in the documentation on my first read through! Thanks for setting me straight.
