### 2022-02-12 FSEC-Photovoltaics/pvrpm-lcoe#31 @blthayer
Maybe I'm confused :smile:

My understanding is that after exporting the SAM case to `.json`, we need to reflect the module order found in the exported `json` into the `.yaml` configuration file under the `module_order` key, correct?

I suppose I don't understand why a user needs to look up the module order in the [gigantic table](https://nrel-pysam.readthedocs.io/en/master/Configs.html#sam-simulation-configurations).

Maybe it'd be a nice feature add for `pvrpm` to parse the `json` and extract the module order itself (*e.g.*, using something similar to the code above)?

Is it ever allowed to put a different module order in the `yaml` config file than what comes out of the exported `json` case file? If not, this manual transcription step should be eliminated in my opinion. Plenty of room for user error :smile:
