### 2022-02-12 FSEC-Photovoltaics/pvrpm-lcoe#36 @blthayer
I'm still getting the same error after updating to use Detailed PV Models and a financial model that supports lifetime costs (as far as I could tell). I've attached a .zip archive with the exported `json` file and `yaml` configuration file
[richland_detailed.zip](https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/files/8054236/richland_detailed.zip).

At any rate, I would consider updating the error message to give the user actionable information.

Looking at [the code](https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/blob/2e8ebc716ca176e304a5f029a2facb792dc68fa3/pvrpm/core/case.py#L87) where I believe the error message originates from, it seems that the module name is being inferred from the name of the `json` file. However, as I mentioned in https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/issues/39, SAM is only exporting a single `json` file for me (and I'm not sure if this is expected behavior or a problem with SAM. As mentioned above SAM does crash after creating the single `json` file plus the `.h` and `.so` files)
