### 2022-02-12 FSEC-Photovoltaics/pvrpm-lcoe#31 @brandons209
The layout of the JSON files isn't the module order. Each JSON file represents a specific module and the order of all of those modules must follow the order in the table as you mentioned. If you put a different module order then what is needed the SAM simulation won't work. I think its possible to add a feature to automatically do this.

Basically, you got a bunch of modules on export to JSON from SAM. Then you would see all the modules you have, find the matching order in the table, then order them properly in the `.yaml` config file.
