### 2022-02-12 FSEC-Photovoltaics/pvrpm-lcoe#36 @brandons209
I will take a look, this is most likely an error in SAM, there should be more then 1 file.
