### 2022-02-12 FSEC-Photovoltaics/pvrpm-lcoe#36 @brandons209
I need you to send me your SAM case file to check it.
