### 2022-02-12 FSEC-Photovoltaics/pvrpm-lcoe#36 @brandons209
Okay, I figured it out. I misread your initial bug report, you are selecting the wrong option when you click the "Generate Code" menu item. You need to do "Generate Code" -> "PySAM JSON" not "JSON for Inputs".
