### 2022-02-15 FSEC-Photovoltaics/pvrpm-lcoe#36 @brandons209
As an addition to this, I realize the `Getting Started` page did say select `JSON for Inputs`, I think this changed with the latest version of SAM. Sorry for the confusion as I didn't realize this, I will fix the docs now.
