### 2022-02-16 bfc44cb
automate calculating module order, fix wrong instructions in getting started, and lists supported financial models
