### 2022-02-16 228d377
fix broken link in getting started
