### 2022-02-19 FSEC-Photovoltaics/pvrpm-lcoe#36 @blthayer
I'm now able to get my case past this error thanks to the updated documentation.
