### 2022-02-23 FSEC-Photovoltaics/pvrpm-lcoe#49 @SarthakJariwala
BUG (maybe): Can't run the simulation
While trying to run pvrpm with a test case and example config file, I am getting the following error - `Illegal instruction while running pvrpm` (see below for more details).

## Error
```console
$ pvrpm run --case test-case-comm example_config.yml --trace

2022-02-22 16:20:21,419--WARNING: There is a non-zero value in the fixed annual O&M costs input. These will be overwritten with the new values.
2022-02-22 16:20:21,419--WARNING: Degradation is set by the PVRPM script, you have entered a non-zero degradation to the degradation input. This script will set the degradation input to zero.
[1]    49994 illegal hardware instruction  pvrpm run --case test-case-comm example_config.yml --trace
```

## Details

I am running everything on 'macOS-10.16-x86_64-i386-64bit'

### Installation
I am using conda to create a new environment and install pvrpm as follows:
```console
conda create -n pvrpm "python>=3.8"
conda activate pvrpm
python -m pip install git+https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/@v1.7.1
```

### SAM
I am using the "Residential Owner" model under "Detailed PV Model" and exporting the PySAM JSON to a folder.

### Config file
I downloaded the example config file and changed the results folder.

### Complete specs of packages
```
# packages in environment at /Users/sjariwala/miniconda3/envs/pvrpm:
#
# Name                    Version                   Build  Channel
bzip2                     1.0.8                h0d85af4_4    conda-forge
ca-certificates           2021.10.8            h033912b_0    conda-forge
certifi                   2021.10.8        py38hecd8cb5_2
click                     8.0.4                    pypi_0    pypi
cycler                    0.11.0                   pypi_0    pypi
fonttools                 4.29.1                   pypi_0    pypi
kiwisolver                1.3.2                    pypi_0    pypi
libcxx                    12.0.1               habf9029_1    conda-forge
libffi                    3.4.2                h0d85af4_5    conda-forge
libzlib                   1.2.11            h9173be1_1013    conda-forge
matplotlib                3.5.1                    pypi_0    pypi
ncurses                   6.3                  he49afe7_0    conda-forge
nrel-pysam                3.0.0                    pypi_0    pypi
nrel-pysam-stubs          3.0.0                    pypi_0    pypi
numpy                     1.22.2                   pypi_0    pypi
openssl                   3.0.0                h0d85af4_2    conda-forge
packaging                 21.3                     pypi_0    pypi
pandas                    1.4.1                    pypi_0    pypi
pillow                    9.0.1                    pypi_0    pypi
pip                       22.0.3             pyhd8ed1ab_0    conda-forge
pvrpm                     1.7.1                    pypi_0    pypi
pyparsing                 3.0.7                    pypi_0    pypi
python                    3.8.12          h43ca1e7_3_cpython    conda-forge
python-dateutil           2.8.2                    pypi_0    pypi
pytz                      2021.3                   pypi_0    pypi
pyyaml                    6.0                      pypi_0    pypi
readline                  8.1                  h05e3726_0    conda-forge
scipy                     1.8.0                    pypi_0    pypi
setuptools                58.0.4           py38hecd8cb5_0
six                       1.16.0                   pypi_0    pypi
sqlite                    3.37.0               h23a322b_0    conda-forge
tk                        8.6.12               h5dbffcc_0    conda-forge
tqdm                      4.62.3                   pypi_0    pypi
tzdata                    2021e                he74cb21_0    conda-forge
wheel                     0.37.1             pyhd8ed1ab_0    conda-forge
xz                        5.2.5                haf1e3a3_1    conda-forge
zlib                      1.2.11            h9173be1_1013    conda-forge
```
