### 2022-02-23 FSEC-Photovoltaics/pvrpm-lcoe#49 @SarthakJariwala
Also, using the `verify` command results in the same error :

```console
$ pvrpm verify --case test-case-comm example_config.yml

2022-02-22 16:43:40,795--WARNING: There is a non-zero value in the fixed annual O&M costs input. These will be overwritten with the new values.
2022-02-22 16:43:40,795--WARNING: Degradation is set by the PVRPM script, you have entered a non-zero degradation to the degradation input. This script will set the degradation input to zero.
[1]    50513 illegal hardware instruction  pvrpm verify --case test-case-comm example_config.yml
```
