### 2022-02-23 FSEC-Photovoltaics/pvrpm-lcoe#50 @SarthakJariwala
DOC : Add documentation for all available commands
Please add documentation (with example usage) for all commands available in `pvrpm`. This will help users quickly understand how to use the different available commands.
