### 2022-02-23 FSEC-Photovoltaics/pvrpm-lcoe#51 @SarthakJariwala
Optional: Support configuration from PySAM (maybe?)
Could you also use pvrpm directly with PySAM?

Instead of creating a case in SAM and exporting it, would it also be possible to develop the case in PySAM and export the configurations to use them with pvrpm? This would eliminate the need for users to download and interact with the GUI every time they want to model a new case.
