### 2022-02-23 FSEC-Photovoltaics/pvrpm-lcoe#52 @SarthakJariwala
DOC: Document public API
The API section in the docs is [empty](https://pvrpm.readthedocs.io/en/latest/api/modules.html). If the API listed there isn't supposed to be public API, I would recommend not listing it there.

As I understand, pvrpm is intended to be used from the command line (and not imported as a module), and as such, commands `run`, `verify` `sim` are the public API. I would suggest documenting those commands in more detail (see #50).
