### 2022-02-23 #4093 @SarthakJariwala
@brandons209 I will use this comment as a checklist for all the issues I have filled in your repository.

- [ ] https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/issues/49
- [ ] https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/issues/50
- [ ] https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/issues/51
- [ ] https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/issues/52
