### 2022-02-23 FSEC-Photovoltaics/pvrpm-lcoe#49 @SarthakJariwala
Maybe this helps in your debugging - I couldn't run the tests as well using pytest (see output below)

```console
$ pytest
=========================================================== test session starts ============================================================
platform darwin -- Python 3.8.12, pytest-7.0.1, pluggy-1.0.0
rootdir: /Users/sjariwala/Documents/JOSS_review/pvrpm-lcoe
collected 5 items

tests/integration/test_simulation.py Fatal Python error: Illegal instruction

Current thread 0x00000001189c8600 (most recent call first):
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/pvrpm/core/case.py", line 763 in value
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/pvrpm/core/case.py", line 540 in __verify_case
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/pvrpm/core/case.py", line 70 in __init__
  File "/Users/sjariwala/Documents/JOSS_review/pvrpm-lcoe/tests/integration/test_simulation.py", line 17 in test_simulation
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/_pytest/python.py", line 192 in pytest_pyfunc_call
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/pluggy/_callers.py", line 39 in _multicall
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/pluggy/_manager.py", line 80 in _hookexec
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/pluggy/_hooks.py", line 265 in __call__
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/_pytest/python.py", line 1718 in runtest
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/_pytest/runner.py", line 168 in pytest_runtest_call
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/pluggy/_callers.py", line 39 in _multicall
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/pluggy/_manager.py", line 80 in _hookexec
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/pluggy/_hooks.py", line 265 in __call__
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/_pytest/runner.py", line 261 in <lambda>
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/_pytest/runner.py", line 340 in from_call
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/_pytest/runner.py", line 260 in call_runtest_hook
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/_pytest/runner.py", line 221 in call_and_report
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/_pytest/runner.py", line 132 in runtestprotocol
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/_pytest/runner.py", line 113 in pytest_runtest_protocol
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/pluggy/_callers.py", line 39 in _multicall
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/pluggy/_manager.py", line 80 in _hookexec
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/pluggy/_hooks.py", line 265 in __call__
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/_pytest/main.py", line 347 in pytest_runtestloop
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/pluggy/_callers.py", line 39 in _multicall
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/pluggy/_manager.py", line 80 in _hookexec
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/pluggy/_hooks.py", line 265 in __call__
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/_pytest/main.py", line 322 in _main
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/_pytest/main.py", line 268 in wrap_session
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/_pytest/main.py", line 315 in pytest_cmdline_main
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/pluggy/_callers.py", line 39 in _multicall
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/pluggy/_manager.py", line 80 in _hookexec
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/pluggy/_hooks.py", line 265 in __call__
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/_pytest/config/__init__.py", line 165 in main
  File "/Users/sjariwala/miniconda3/envs/pvrpm/lib/python3.8/site-packages/_pytest/config/__init__.py", line 188 in console_main
  File "/Users/sjariwala/miniconda3/envs/pvrpm/bin/pytest", line 8 in <module>
[1]    51082 illegal hardware instruction  pytest
```
