### 2022-02-23 FSEC-Photovoltaics/pvrpm-lcoe#51 @brandons209
You can, following PySAM's documentation. However their documentation recommends to use the SAM interface as its tedious to configure every variable needed manually for a system configuration, compared to exporting the JSON files from SAM.

https://nrel-pysam.readthedocs.io/en/master/Examples.html

As for support in PVRPM, a user can write their own Python script to generate the JSONs needed for PVRPM to work, but it seems to be easier to use the GUI. I also think it opens up for a lot more user error compared to using the GUI. I think it would be possible to add in the functionality of PVRPM being able to take in prebuilt module objects but it would be non-trivial.
