### 2022-02-23 FSEC-Photovoltaics/pvrpm-lcoe#52 @brandons209
Huh, odd, was showing up fine locally. I'll fix it. Also, PVRPM can be used from both the command line and imported as modules. to manually setup and run a case with your own wrapper scripts.
