### 2022-02-23 FSEC-Photovoltaics/pvrpm-lcoe#49 @brandons209
Some quick googling suggests this is a macOS specific issue with running Python on it, more specifically to your processor. My guess is that one of the packages PVRPM uses is not supported for your CPU architecture, even though I think you are using an Intel x86_64 architecture processor, although I am not familiar with mac products. If you have the M1 chip that may be the issue. Unfortunately, I do not have easy access to a macOS device for testing. The error when running verify suggests Python itself is segfaulting with that error, which various articles on the Google search seem to confirm. I may be able to get a hold of a macOS device for testing but I am not sure right now, may have to try to setup a VM, but a VM environment may not produce the specific issue you are having.

My best guess is that is probably pandas or numpy causing the issue, so maybe finding and installing mac specific versions for those packages or researching issues along with those packages on macOS? I'll try to do some digging and post anything relevant I can find here. It may be that PVRPM just won't work on some macOS computers.
