### 2022-02-23 488827b
fix some issues with config, simulation. added usage to docs
