+++ docs/website/_sources/tutorial_assumptions_limitations.rst.txt
@@ -1,8 +1,0 @@
-SAM/PVRPM Assumptions / Limitations
-=====================================
-.. toctree::
-  :hidden:
-
-PVRPM makes a few assumptions in order to be able to run a realistic simulation efficiently.
-
-TODO
