+++ docs/website/_sources/api/pvrpm.core.modules.rst.txt
@@ -1,37 +1,0 @@
-pvrpm.core.modules package
-==========================
-
-Submodules
-----------
-
-pvrpm.core.modules.failure module
----------------------------------
-
-.. automodule:: pvrpm.core.modules.failure
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pvrpm.core.modules.monitor module
----------------------------------
-
-.. automodule:: pvrpm.core.modules.monitor
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pvrpm.core.modules.repair module
---------------------------------
-
-.. automodule:: pvrpm.core.modules.repair
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: pvrpm.core.modules
-   :members:
-   :undoc-members:
-   :show-inheritance:
