+++ docs/website/_sources/example_pvrpm_config.rst.txt
@@ -1,9 +1,0 @@
-Example YAML Configuration
-===========================
-.. toctree::
-  :hidden:
-
-Please read all the comments carefully, there are many options for each section of PVRPM.
-
-.. include:: ../pvrpm/config/example_config.yml
-  :code: yaml
