+++ docs/website/_sources/tutorial_getting_started.rst.txt
@@ -1,341 +1,0 @@
-Getting Started
-=================================
-.. toctree::
-  :hidden:
-
-PVRPM enhances SAM's models for PV systems to obtain a more accurate LCOE than its base simulation. It also allows studying the effects of monitoring and repair techniques on PV systems throughout their lifetime.
-
-To get started, create a new SAM case in the SAM GUI. From there, you **must choose the Detailed Photovoltaic model** for PVRPM to work. This is required because PVRPM need's specific parameters that only exist in this model.
-
-Then, choose your financial model. It must be a financial model that supports lifetime losses. Please read SAM's documentation for help in setting up the case.
-
-Once that is set up, you can download the example configuration and modify it as needed. Below explains from start to finish of how to run a simulation with PVRPM.
-
-Exporting the SAM Case
---------------------------------
-PVRPM works by reading the information in your SAM case and the PVRPM YAML configuration to properly run the simulation. For the first step, you need to export your SAM case to JSON files, which PVRPM uses to read in your SAM case.
-
-To do this:
-1. Open the SAM GUI, then open your `.sam` case file.
-2. Once it is open, click the drop-down menu for the case, which is located next to the case name on the top bar.
-3. Click `Generate Code`, then `JSON for inputs`.
-4. A file explorer window will open, allowing you to select a place to save the JSON files. Select an empty folder.
-
-Once this is done, the selected folder will contain a few JSON files. The names of these files will follow the convention of `case-name_module.json` where `case-name` is the name of the case, and `module` is the module that JSON represents. Pay attention to what modules you have; you'll need to know that for the next part. You can remove the `.h` and `.so` files.
-
-Configuring PVRPM
-------------------------
-
-This will go over every configuration option to set up the case study step by step. The example configuration is also heavily commented on to help with the parameters. Also, please study the logic diagram as it can help when setting up the configuration here. *Also note all of the values listed in examples are entirely arbitrary and do not represent a realistic case.*
-
-Module Order
-----------------
-These modules correspond to the modules listed in the JSON files you obtained. The modules must be simulated in the correct order for PVRPM to properly run simulations using PySAM (SAM's Python interface). A typical order is in the example, but you may need to modify it depending on your case.
-
-To make the module order easy to set up, go to this website: https://nrel-pysam.readthedocs.io/en/master/Configs.html#sam-simulation-configurations
-
-Then, find the modules you have under `SSC Compute Modules` column in the table; the SAM configuration column should be one of the `Detailed PV Model` rows. Once you find one containing all your modules, put them in the order as they appear in the `SSC Compute Modules` column.
-
-.. code-block:: yaml
-  :linenos:
-
-  module_order: # the order the modules of this case should be executed
-                # check to see what modules you use by the name of the json output files
-                # typically, this should be Pvsamv1 -> grid -> utiltiy -> others
-                # see https://nrel-pysam.readthedocs.io/en/master/Configs.html#sam-simulation-configurations
-  # Only detailed PV models are supported, which is Pvsamv1, your case must have this module
-  # Also, only LCOE calculators that support lifetime are allowed, PVRPM will check this and notify you if your LCOE calculator doesn't support lifetime
-    - Pvsamv1
-    - Grid
-    - Utilityrate5
-    - Cashloan
-
-
-Run Setup
-----------------
-Here, set the results folder location. On Windows, use only one backslash to separate directories; you do not need to escape them or spaces. Then set the number of realizations you want to run and the confidence interval for calculating results.
-
-The results folder and number of realizations can be overridden from the command line.
-
-.. code-block:: yaml
-  :linenos:
-
-  ### Run Setup ###
-  # results folder doesn't need to exist, it will be created if it doesnt already
[…]
