### 2022-02-23 a098c23
fix some issues with config, simulation. added usage to docs
