+++ docs/tutorial_1installation.rst
@@ -1,5 +1,7 @@
 Installation
 =================================
+.. toctree::
+  :hidden:
 
 This document covers installation and setup of the tool. The tool requires you to build a valid case in SAM, so if you haven't already download and install SAM from here: https://sam.nrel.gov/download.html
 
