+++ docs/tutorial_assumptions_limitations.rst
@@ -1,7 +1,5 @@
 PVRPM Assumptions / Limitations
 =====================================
-.. toctree::
-  :hidden:
 
 PVRPM makes a few assumptions in order to be able to run a realistic simulation efficiently. Along with these assumptions arise limitations to what the model can simulate realistically.
 
