+++ .readthedocs.yaml
@@ -2,3 +2,8 @@ version: 2
 
 python:
   version: 3.8
+  install:
+    - method: pip
+      path: .
+      extra_requirements:
+        - docs
