### 2022-02-23 FSEC-Photovoltaics/pvrpm-lcoe#50 @brandons209
Fixed in a098c23e4f051e1a9f7099464984b2ec3331c258
