### 2022-02-23 FSEC-Photovoltaics/pvrpm-lcoe#52 @brandons209
Fixed in 4899e72e970b94dcf715e760ba2e0a2aebf553a2
