### 2022-02-23 FSEC-Photovoltaics/pvrpm-lcoe#49 @SarthakJariwala
I am running it on a Mac with Intel chip.

I don't think it is related to numpy or pandas. I just tried running some NumPy and pandas-related computation in the pvrpm environment I created, and everything works as it should have. I have also been using NumPy and pandas in other environments without any issues.

One thing I would suggest is to add macOS to your test workflow using github actions and see if the tests pass/run. It might be faster than getting hold of a mac.
(Along those lines, I see you are already testing on ubuntu but I would also recommend adding macOS and windows to your test workflow)
