### 2022-02-23 FSEC-Photovoltaics/pvrpm-lcoe#49 @brandons209
Alright, I didn't realize actions can test on multiple OSes. I'll go ahead and add them and see what happens. If its not those packages then something is causing Python itself to segfault, which would be hard to diagnose.
