### 2022-02-23 FSEC-Photovoltaics/pvrpm-lcoe#51 @SarthakJariwala
Ok, thanks for the information. Having worked with PySAM before, I agree with you; it is pretty tedious to set up the simulation.
