### 2022-02-24 0fdbea3
fix some issues with output generationg and pytest
