### 2022-02-24 FSEC-Photovoltaics/pvrpm-lcoe#49 @brandons209
Alright, so the action fails on macOS. It has to be one of the packages causing the issue. Could you do me a favor?

Can you import each library PVRPM uses one at a time and see if any of them cause the crash?
Just run each code snippet one at a time in a interpreter and see if any cause a crash.

```
import PySAM.PySSC as pssc
ssc = pssc.PySSC()
```

Loading all PySAM packages:
```
def load_pysam_modules():
    global pysam
    import PySAM as pysam

    for loader, module_name, is_pkg in pkgutil.walk_packages(pysam.__path__):
        try:
            print(f"Loading: {module_name}")
            importlib.import_module(f"{pysam.__name__}.{module_name}")
        except:
            pass

load_pysam_modules()
```

```
import scipy
import scipy.stats as stats
from scipy.special import gamma, gammaln
```

`import click`

`import yaml`

`import matplotlib.pyplot as plt`

`from dateutil.relativedelta import relativedelta`

`from tqdm import tqdm`

```
import numpy as np
import pandas as pd
```
