### 2022-02-24 FSEC-Photovoltaics/pvrpm-lcoe#49 @brandons209
Hopefully one of those will cause the crash and I can see if there is a workaround for that package.
