### 2022-02-24 FSEC-Photovoltaics/pvrpm-lcoe#49 @SarthakJariwala
I changed the `except: pass` to print the `Exception` in PySAM snippet and got the following Exceptions (complete output is underneath that) -

All other snippets worked.

```console
dynamic module does not define module export function (PyInit_libSAM_api)
dynamic module does not define module export function (PyInit_libssc)
No module named 'requests'
```

Complete output from PySAM snippet:
```console
Loading : AdjustmentFactors
Loading : Annualoutput
Loading : Battery
Loading : BatteryStateful
Loading : BatteryTools
Loading : Battwatts
Loading : Belpe
Loading : Biomass
Loading : Cashloan
Loading : CbConstructionFinancing
Loading : CbEmpiricalHceHeatLoss
Loading : CbMsptSystemCosts
Loading : Communitysolar
Loading : DsgFluxPreprocess
Loading : Equpartflip
Loading : EtesElectricResistance
Loading : Fuelcell
Loading : GenericSystem
Loading : Geothermal
Loading : GeothermalCosts
Loading : Grid
Loading : Hcpv
Loading : HostDeveloper
Loading : Iec61853interp
Loading : Iec61853par
Loading : InvCecCg
Loading : IphToLcoefcr
Loading : Ippppa
Loading : Irradproc
Loading : IsccDesignPoint
Loading : Layoutarea
Loading : Lcoefcr
Loading : Levpartflip
Loading : LinearFresnelDsgIph
Loading : Merchantplant
Loading : MhkCosts
Loading : MhkTidal
Loading : MhkWave
Loading : Poacalib
Loading : Pv6parmod
Loading : PvGetShadeLossMpp
Loading : Pvsamv1
Loading : Pvsandiainv
Loading : Pvwattsv1
Loading : Pvwattsv11ts
Loading : Pvwattsv1Poa
Loading : Pvwattsv5
Loading : Pvwattsv51ts
Loading : Pvwattsv7
Loading : Pvwattsv8
Loading : PySSC
Loading : ResourceTools
No module named 'requests'
Loading : Saleleaseback
Loading : Sco2AirCooler
Loading : Sco2CompCurves
Loading : Sco2CspSystem
Loading : Sco2CspUdPcTables
Loading : Singlediode
Loading : Singlediodeparams
Loading : Singleowner
Loading : SixParsolve
Loading : Snowmodel
Loading : Solarpilot
Loading : Swh
Loading : TcsMSLF
Loading : TcsgenericSolar
Loading : TcslinearFresnel
Loading : TcsmoltenSalt
Loading : TcstroughEmpirical
Loading : TcstroughPhysical
Loading : TestUdPowerCycle
Loading : Thermalrate
Loading : Thirdpartyownership
Loading : Timeseq
Loading : TroughPhysical
Loading : TroughPhysicalCspSolver
Loading : TroughPhysicalProcessHeat
Loading : UiTesCalcs
Loading : UiUdpcChecks
Loading : UserHtfComparison
Loading : Utilityrate
Loading : Utilityrate2
Loading : Utilityrate3
Loading : Utilityrate4
Loading : Utilityrate5
Loading : WaveFileReader
Loading : Wfcheck
Loading : Wfcsvconv
Loading : Wfreader
Loading : WindFileReader
Loading : WindObos
Loading : Windbos
Loading : Windcsm
Loading : Windpower
Loading : libSAM_api
dynamic module does not define module export function (PyInit_libSAM_api)
Loading : libssc
dynamic module does not define module export function (PyInit_libssc)
Loading : version
Loading : version_differ
No module named 'requests'
```
