### 2022-02-24 FSEC-Photovoltaics/pvrpm-lcoe#49 @brandons209
Those are fine, but every snippet worked? That's really odd then...

If verify causes a crash then we can try narrowing it down a bit. Its not with the SAM simulation or PVRPM simulation.

Its crashing somewhere in creating the SAM case in PVRPM.

If you can clone the repo to your local computer, then cd into the repo and do `pip install -e .`

Then go to `pvrpm > core > case.py` and add in print statements after every line in the __init__ function to try to narrow down where this crash occurs.

Something like:
```
def __init__(self, sam_json_dir: str, config: str, num_realizations: int = 0, results_folder: str = None):
        """"""
        print("enter init")
        self.ssc = pssc.PySSC()
        print("test 1")
        self.config = self.__load_config(config, type="yaml")
        print("test 2")
        self.sam_json_dir = sam_json_dir
        self.daily_tracker_coeffs = None
        self.modules = self.__load_modules()
        print("test 3")

        # will be calculated after base case simulation
        self.daylight_hours = None
        self.annual_daylight_hours = None
        self.base_lcoe = None
        self.base_npv = None
        self.base_ac_energy = None
        self.base_annual_energy = None
        self.base_dc_energy = None

        self.base_load = None
        self.base_tax_cash_flow = None
        self.base_losses = {}

        if not (self.modules and self.config):
            raise CaseError("There are errors in the configuration files, see logs.")

        # override results folder and number of realizations
        print("test 4")
        if num_realizations >= 2:
            self.config[ck.NUM_REALIZATION] = num_realizations
        if results_folder is not None:
            self.config[ck.RESULTS_FOLDER] = results_folder

        # lookup table for module order:
        self.module_orders = [
            ["Pvsamv1", "Grid", "Utilityrate5", "Cashloan"],
            ["Pvsamv1", "Grid", "Utilityrate5", "Merchantplant"],
            ["Pvsamv1", "Grid", "Utilityrate5", "Levpartflip"],
            ["Pvsamv1", "Grid", "Utilityrate5", "Equpartflip"],
            ["Belpe", "Pvsamv1", "Grid", "Utilityrate5", "Cashloan"],
            ["Pvsamv1", "Grid", "Utilityrate5", "Saleleaseback"],
            ["Pvsamv1", "Grid", "Utilityrate5", "Singleowner"],
            ["Pvsamv1", "Grid", "Utilityrate5", "HostDeveloper"],
            ["Belpe", "Pvsamv1", "Grid", "Utilityrate5", "Thirdpartyownership"],
        ]

        # lookup table for models that pvrpm cannot use
        self.bad_module_orders = [
            ["Pvsamv1", "Grid"],
            ["Pvsamv1", "Grid", "Lcoefcr"],
        ]
        print("test 5")
        self.__verify_case()
        print("test 6")
        self.__verify_config()
        print("test 7")
```

Just try and narrow down where this is happening. You can also try using the debugger but I haven't had much experience with the Python debugger, if you do and want to try that please go ahead.
