### 2022-02-27 b8a9221
added some information about the base case simulation in assumptions
