+++ docs/tutorial_assumptions_limitations.rst
@@ -13,4 +13,6 @@ Availability is also based upon the daylight hours for the configured location v
 
 During the calculation of when repairs, monitoring, and failures take place, it is assumed that as soon as a failure occurs, either monitoring or repair immediately begins the next day, depending on what is configured for that component. This must be accounted for in the user's configuration of the distributions. If multiple monitoring types are defined for a single component, then the quickest monitoring will override the others. For example, if component-level monitoring is defined for modules and cross-level monitoring from the inverter to the module, whatever has the shortest time to detection. Typically, it would be the component level monitoring.
 
+With the configuration and outputs for the realizations, PVRPM also compares the ``base case`` in its outputs. The ``base case`` is what the system would output if there were no failures for any component on all of the component levels. The ``base case`` only has module degradation throughout the system's lifetime, depending on the configuration (which could be none if no module degradation is configured). For modeling small systems with module degradation and failures, you might see that monthly energy production is lower for the ``base case`` than the realizations since modules perform much more poorly in the later years of the simulation. Modules in the realizations will fail and be repaired, so their degradation will reset and operate at peak efficiency. This constitutes an overall higher monthly energy production across the lifetime than the base case. This is expected behavior for this scenario.
+
 Finally, each realization is independent of the other realizations in a simulation run. A realization is simply a single "run" of the system for the duration of its lifetime, with the provided configurations.
