### 2022-03-01 FSEC-Photovoltaics/pvrpm-lcoe#49 @brandons209
If you don't have the time or unable I can try to do it with the github action as well, just let me know.
