+++ .github/ISSUE_TEMPLATE/question.md
@@ -1,0 +1,23 @@
+---
+name: Question
+about: Ask a question and get support
+title: "[QUESTION] <please fill with brief question title>"
+labels: question
+assignees: brandons209
+
+---
+
+**Ask the question**
+Provide a brief and concise question.
+
+**Where have I looked for the answer?**
+List out where you looked for the answer, [e.g. I checked the installation documentation, assumptions, readme, other GitHub issues, etc]
+
+**Describe the question**
+Please provide as much detail as possible in what you are asking. If applicable include screenshots, references to other issues, an example that relates to the question, etc.
+
+**Provide SAM case and PVRPM configuration**
+If your question relates to running a PVRPM simulation, please provide a zip file or a link to a zip file containing:
+  - PVRPM `yaml` configuration file
+  - JSON files representing the SAM case (exported from SAM)
+  - Weather file used in the SAM case (location can be found by looking at the weather file in the SAM's GUI for the case)
