+++ .github/ISSUE_TEMPLATE/documentation-change.md
@@ -1,0 +1,16 @@
+---
+name: Documentation Change
+about: Request a change to documentation
+title: "[DOCs] <title of doc change here>"
+labels: documentation
+assignees: ''
+
+---
+
+**Provide a link to the documentation page that needs edits**
+Can be linked to the readthedocs site or documentation file in the GitHub repository. If requesting a new page, list the page name and description of what the page should contain.
+
+**Describe the changes**
+Can be generic as "this needs to be clearer" or specific changes like "the grammar here is incorrect, change it to ..."
+
+If wanting to add a new page, describe what the new page is and why it should be added.
