+++ .github/ISSUE_TEMPLATE/feature-request.md
@@ -1,0 +1,21 @@
+---
+name: Feature Request
+about: Suggest an enhancement to this project
+title: "[FEATURE] <replace with feature title>"
+labels: enhancement
+assignees: ''
+
+---
+
+**Is your feature request related to a problem? Please describe.**
+A clear and concise description of what the problem is. Ex. I'm always frustrated when [...]
+Please link any relevant issues or commits to this feature request.
+
+**Describe the solution you'd like**
+A clear and concise description of what you want to happen.
+
+**Provide an example of the feature**
+Could be an example output result file, example of the configuration file, analogy to a real world phenomena in PV systems, etc.
+
+**Additional context**
+Add any other context or screenshots about the feature request here.
