+++ .github/ISSUE_TEMPLATE/bug-report.md
@@ -1,0 +1,40 @@
+---
+name: Bug Report
+about: Submit a bug in the PVRPM package
+title: "[BUG] <please replace this with bug title>"
+labels: bug
+assignees: brandons209
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To Reproduce**
+Steps to reproduce the behavior:
+1. Run command '..'
+2. See error
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Case file**
+Please include a zip file with the issue submission, or link to a zip file containing:
+  - PVRPM `yaml` configuration file
+  - JSON files representing the SAM case (exported from SAM)
+  - Weather file used in the SAM case (location can be found by looking at the weather file in the SAM's GUI for the case)
+
+**System Information (please complete the following information):**
+ - OS: [e.g. macOS 10.16, Arch Linux, Windows 10, Ubuntu 20.04]
+ - CPU: [e.g. AMD Ryzen 7 2700x] 
+ - Python Version: [e.g. 3.8, 3.9, 3.10]
+ - PVRPM Version: [e.g. 1.7.1, 1.7.2]
+
+**Stack Trace**
+If applicable, provide a full stack trace here. Use the `--trace` flag when running a PVRPM simulation to print stack traces.
+```
+<stack trace>
+```
+
+**Additional context**
+Add any other context about the problem here.
