+++ setup.py
@@ -1,4 +1,11 @@
 from setuptools import setup, find_packages
+import os
+
+
+def read_file(filename):
+    with open(os.path.join(os.path.dirname(__file__), filename)) as file:
+        return file.read()
+
 
 setup(
     name="pvrpm",
@@ -15,23 +22,14 @@ setup(
         "matplotlib",
         "python-dateutil",
     ],
-    extras_require={
-        "docs": [
-            "sphinx",
-            "sphinx_rtd_theme",
-            "sphinxcontrib-napoleon",
-        ],
-        "testing": [
-            "pytest",
-        ],
-    },
+    extras_require={"docs": ["sphinx", "sphinx_rtd_theme", "sphinxcontrib-napoleon",], "testing": ["pytest",],},
     python_requires=">=3.8",
     entry_points="""
         [console_scripts]
         pvrpm=pvrpm.__main__:main
     """,
     author=["Brandon Silva", "Paul Lunis"],
-    long_description="README.md",
+    long_description=read_file("README.md"),
     long_description_content_type="text/markdown",
     url="https://github.com/FSEC-Photovoltaics/pvrpm-lcoe",
 )
