### 2022-03-02 FSEC-Photovoltaics/pvrpm-lcoe#49 @SarthakJariwala
I won't be able to give it a try until the weekend. Please go ahead and try it with GitHub Actions.
