### 2022-03-02 f1d0401
figure out where macos is crashing using action
