+++ .github/workflows/tests.yml
@@ -6,28 +6,8 @@ on:
   push:
 
 jobs:
-  pytest-linux:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-python@v2
-        with:
-          python-version: "3.8.x"
-          architecture: "x64"
-      - run: pip install .[testing] && pytest
-
   pytest-macos:
     runs-on: macos-latest
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-python@v2
-        with:
-          python-version: "3.8.x"
-          architecture: "x64"
-      - run: pip install .[testing] && pytest
-
-  pytest-windows:
-    runs-on: windows-latest
     steps:
       - uses: actions/checkout@v2
       - uses: actions/setup-python@v2
