### 2022-03-02 f3fbeb8
stop capture on pytest output
