+++ .github/workflows/tests.yml
@@ -14,4 +14,4 @@ jobs:
         with:
           python-version: "3.8.x"
           architecture: "x64"
-      - run: pip install .[testing] && pytest
+      - run: pip install .[testing] && pytest -s
