+++ pvrpm/core/case.py
@@ -154,6 +154,7 @@ class SamCase:
         """
         Verifies loaded YAML configuration file.
         """
+        logger.warning("test 18")
         # helper function to check distribution parameters
         def check_params(component: str, name: str, config: dict):
             if ck.MEAN not in config[ck.PARAM]:
@@ -165,43 +166,47 @@ class SamCase:
             elif config[ck.DIST] != ck.EXPON and ck.STD not in config[ck.PARAM]:
                 raise CaseError(f"STD parameter for {name} is missing!")
 
+        logger.warning("test 19")
         # tracking check
         if self.config[ck.TRACKING] and not self.config.get(ck.TRACKER, False):
             raise CaseError(
                 "Tracking modules loaded in SAM config but no trackers defined in the PVRPM YML configuration, please make sure it is setup!"
             )
-
+        logger.warning("test 20")
         if not self.config[ck.TRACKING] and self.config.get(ck.TRACKER, False):
             logger.warning(
                 "No tracking modules loaded for the SAM case, however there is a tracking configuration in the PVRPM configuration. This configuration will be ignored and no simulation with trackers will be performed."
             )
 
+        logger.warning("test 21")
         top_level_keys = set(ck.needed_keys)
         included_keys = set(self.config.keys()) & top_level_keys
         if included_keys != top_level_keys:
             raise CaseError(
                 f"Missing required configuration options in the PVRPM configuration: {top_level_keys - included_keys}"
             )
-
+        logger.warning("test 22")
         if self.config[ck.NUM_REALIZATION] < 2:
             raise CaseError("Number of realizations must be greater than or equal to 2!")
-
+        logger.warning("test 23")
         if self.config[ck.TRACKING] and self.config[ck.NUM_TRACKERS] <= 0:
             raise CaseError("If tracking is defined the number of trackers must be greater than 0!")
-
+        logger.warning("test 24")
         if self.config[ck.NUM_COMBINERS] <= 0:
             raise CaseError("Number of combiners must be greater than 0!")
-
+        logger.warning("test 25")
         # static monitoring
         if self.config.get(ck.INDEP_MONITOR, None):
             needed_keys = set(ck.indep_monitor_keys)
             for name, monitor_config in self.config[ck.INDEP_MONITOR].items():
+                logger.warning("test 26")
                 included_keys = set(monitor_config.keys()) & needed_keys
                 unknown_keys = (
                     set(monitor_config.keys())
                     - needed_keys
                     - {ck.INTERVAL, ck.FAIL_THRESH, ck.DIST, ck.PARAM, ck.FAIL_PER_THRESH}
                 )
+                logger.warning("test 27")
                 if included_keys != needed_keys:
                     raise CaseError(f"Independent monitoring for {name} is missing keys {needed_keys - included_keys}")
                 if ck.INTERVAL not in monitor_config and (
@@ -218,7 +223,7 @@ class SamCase:
                         )
                     if monitor_config[ck.DIST] in ck.dists:
                         check_params("", name, monitor_config)
-
+                logger.warning("test 28")
                 if unknown_keys:
                     logger.warning(f"Unknown keys in independent monitoring configuration: {unknown_keys}")
                 for level in monitor_config[ck.LEVELS]:
@@ -228,19 +233,23 @@ class SamCase:
                         self.config[level][ck.INDEP_MONITOR][name] = monitor_config[ck.INTERVAL]
                     else:
                         self.config[level][ck.INDEP_MONITOR][name] = None
-
+        logger.warning("test 29")
         # cross level monitoring and compounding
         if self.config.get(ck.COMP_MONITOR, None):
             # parse levels in order from lowest -> highest to maintain priority on monitoring
             for component in ck.compound_levels:
[…]
