### 2022-03-02 FSEC-Photovoltaics/pvrpm-lcoe#49 @brandons209
I have no clue how to even tackle this. It seems adding in logging statements around the lines that cause this issue seems to prevent the issue from occuring. It happened in a few different spots but I think its occurring when trying to get a value from the PySAM modules.
