### 2022-03-02 FSEC-Photovoltaics/pvrpm-lcoe#49 @brandons209
Could you try running the tests for PySAM itself? I suspect that might help reveal the issue. The repo is here: https://github.com/NREL/pysam
