### 2022-03-03 FSEC-Photovoltaics/pvrpm-lcoe#49 @SarthakJariwala
> It makes no sense that adding in a logging statement before and after those lines makes them run fine.

Yeah, that's very odd! Can you push those changes to main or another branch, and I can try it out whenever I get some time?

---
> Could you try running the tests for PySAM itself?

I ran PySAM tests on my mac, and they passed.

I also tried re-running PVRPM tests with pysam <3.0 (thinking something went wrong with pysam 3.0) and still ran into the same errors.
I also uninstalled everything and re-installed pysam from conda but still ran into the same issues. (I thought something might not have gone right with `pip install` - it's known to create problems sometimes).

---
> I think maybe those who use PySAM don't use macOS and that it just doesn't work on that OS.

I have used PySAM on macOS in the past, and it worked without any issues.
