### 2022-03-03 FSEC-Photovoltaics/pvrpm-lcoe#49 @brandons209
Yeah the branch I am using is called `macos-test` you can pull it and modify it.

I really have no clue then on why that's occurring then. It seems the simulations work but not pulling a value from the modules, but then it works when putting a logging statement before and after the line.

If you have some old code using PySAM, could you try using the `value` method on one of the modules to pull out values? Cause that seems to be where the issue is occurring in PVRPM.
