### 2022-03-03 FSEC-Photovoltaics/pvrpm-lcoe#49 @brandons209
Maybe it is because a value doesn't exist in a PySAM module and it crashes there? In PVRPM I just made a generic function for getting values that may be what's causing the issue too.
