### 2022-03-06 FSEC-Photovoltaics/pvrpm-lcoe#58 @blthayer
[FEATURE] y-axis label for LCOE Box Plot
Looking through the results, all plots except the "LCOE Box Plot" have appropriate y-axis labels.
![LCOE Box Plot](https://user-images.githubusercontent.com/14652512/156938329-1cc2f548-f496-4648-9a3d-ea1b989c94d9.png)
