### 2022-03-06 FSEC-Photovoltaics/pvrpm-lcoe#60 @blthayer
[DOCs] JOSS Paper Feedback
**Provide a link to the documentation page that needs edits**
[Paper link](https://github.com/openjournals/joss-papers/blob/joss.04093/joss.04093/10.21105.joss.04093.pdf)

**Describe the changes**

- Consider adding the word "solar" to the first sentence of the paper, e.g. "... analysis of _solar_ photovoltaic (PV)..." JOSS suggests the high level summary should be suitable "for a diverse, non-specialist audience," and in my experience, you'd be surprised how many people don't know the term "photovoltaic." Also having "solar" in there will help people find your paper/software.
- "LK-based": acronym without spelling it out. Please spell out first. I had to Google this one :)
- Third sentence reads "It allows the user to define..." - it isn't clear if "it" refers to `PVRPM` or `PyPVRPM`
- (OPTIONAL) Fourth sentence (starts with "PyPVRMP is a simulation tool") - consider starting the sentence with something like "By contrast," or something similar to improve flow and explicitly indicate "these things are different in PyPVRPM compared to PVRPM"
- "O&M": acronym without spelling it out. Please spell out. I didn't have to Google this one ;)

Nice work in keeping the write-up concise while getting the important information across!
