### 2022-03-06 #4093 @blthayer
@brandons209 - We're close! I've added three issues (one optional) to my [issue-tracking-checklist](https://github.com/openjournals/joss-reviews/issues/4093#issuecomment-1030686522). After these are addressed, I will consider my review complete.

Nice work! I hope that between mine and Sarthak's review that you feel that `PyPVRPM` has been made better and more approachable.

My general advice for you going forward:

Adding a real-world case study to your documentation could be useful for a newcomer understanding the what and why of `PyPVRPM`. For example, something like "PyPVRPM was used to analyze a PV installation with a capacity of X to evaluate if using infrared drone monitoring was a cost effective preventative maintenance strategy. It was found that Y more MWh would be produced over the plant's lifetime for a total cost savings of Z."

In my experience, good software development practices aren't something taught in school (at least not in electrical engineering), and these skills are something you have to pick up and hone on your own. I didn't read the source code in depth, but I did glance at your testing suite. For starters, I want to say it's _excellent_ that `a)` you have tests and that `b)` you have a full end-to-end simulation test. However, I found the unit tests to be lacking. Learning to write good tests and testable software takes time, but it's well worth the effort. Consider trying to break the functionality of `PyPVRPM` into smaller, easily testable pieces. Using a tool like [pytest-cov](https://pypi.org/project/pytest-cov/) can help you find untested areas of your software (always shoot for 100% coverage! :smile: ). Once you have a comprehensive test suite put together, you can feel really confident that the addition of a new feature or a bug fix didn't break anything or cause regressions. Additionally, rigorous testing helps ensure that not only does the software run, but that it is _correct_. All too often I see academic research software with little to no tests, and in my opinion, untested software is incorrect, untrustworthy software. Alright, putting my testing soap box away now :smile:

I hope that industry and researchers alike pick up `PyPVRPM` and find it useful!

@jgostick - once I'm satisfied that Mr. Silva has addressed https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/issues/58 and https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/issues/60, I'll consider my review complete, and will say as much in a new comment in this thread.
