### 2022-03-06 #4093 @brandons209
> Adding a real-world case study to your documentation could be useful for a newcomer understanding the what and why of PyPVRPM. For example, something like "PyPVRPM was used to analyze a PV installation with a capacity of X to evaluate if using infrared drone monitoring was a cost effective preventative maintenance strategy. It was found that Y more MWh would be produced over the plant's lifetime for a total cost savings of Z."

We are planning to add a real world case, once we finish up a paper that uses PyPVRPM for showing how monitoring effects LCOE. The paper should be done in the next month or two (hopefully).

>In my experience, good software development practices aren't something taught in school (at least not in electrical engineering), and these skills are something you have to pick up and hone on your own. I didn't read the source code in depth, but I did glance at your testing suite. ...

Thank you for the kind words! I do agree that in academia documentation and good software practices aren't used that often like for industry products (and even a lot of industry products too!).
I also agree that comment about testing, but for PVRPM I was having trouble figuring out what to add for unit tests. Everything is tested with the simulation test, so I wasn't sure what else to add. I'll check out that package though!
