### 2022-03-06 FSEC-Photovoltaics/pvrpm-lcoe#58 @brandons209
Will do, also I am a bit worried if your LCOE is in the thousands. Definitely should not be that high.
