+++ paper.md
@@ -30,11 +30,11 @@ bibliography: paper.bib
 
 # Summary
 
-The ability to perform accurate techno-economic analysis of photovoltaic (PV) systems is essential for bankability and investment purposes. Most energy yield models assume an almost flawless operation (i.e., no failures); however, realistically, components fail and get repaired stochastically. This package, PyPVRPM, is a Python translation and improvement of the LK-based PhotoVoltaic Reliability Performance Model (PVRPM), which was first developed at Sandia National Laboratories in Goldsim software [@PVRPM:2011] [@PVRPM:2012]. It allows the user to define a PV system at a specific location and incorporate failure, repair, and detection rates and distributions to calculate energy yield and other financial metrics such as the levelized cost of energy and net present value [@PVRPM:2017]. PyPVRPM is a simulation tool that uses NREL's Python interface for System Advisor Model (SAM) [@SAM:2020] [@PYSAM] to evaluate the performance of a PV plant throughout its lifetime by considering component reliability metrics. Besides the numerous benefits from migrating to Python (e.g., speed, libraries, batch analyses, etc.), it also expands on the failure and repair processes from the LK version by including the ability to vary monitoring strategies. These failures, repairs, and monitoring processes are based on user-defined distributions and values, enabling a more accurate and realistic representation of cost and availability throughout a PV system's lifetime.  
+The ability to perform accurate techno-economic analysis of solar photovoltaic (PV) systems is essential for bankability and investment purposes. Most energy yield models assume an almost flawless operation (i.e., no failures); however, realistically, components fail and get repaired stochastically. This package, PyPVRPM, is a Python translation and improvement of the Language Kit (LK) based PhotoVoltaic Reliability Performance Model (PVRPM), which was first developed at Sandia National Laboratories in Goldsim software [@PVRPM:2011] [@PVRPM:2012]. PyPVRPM allows the user to define a PV system at a specific location and incorporate failure, repair, and detection rates and distributions to calculate energy yield and other financial metrics such as the levelized cost of energy and net present value [@PVRPM:2017]. Our package is a simulation tool that uses NREL's Python interface for System Advisor Model (SAM) [@SAM:2020] [@PYSAM] to evaluate the performance of a PV plant throughout its lifetime by considering component reliability metrics. Besides the numerous benefits from migrating to Python (e.g., speed, libraries, batch analyses, etc.), it also expands on the failure and repair processes from the LK version by including the ability to vary monitoring strategies. These failures, repairs, and monitoring processes are based on user-defined distributions and values, enabling a more accurate and realistic representation of cost and availability throughout a PV system's lifetime.  
 
 # Statement of need
 
-As photovoltaic technology becomes cheaper and more widely used, solutions are needed to accurately assess their performance and monitor them to safeguard performance and reduce downtime. Existing PV performance models assume ideal operation without considering any failures or repairs. In contrast, the value of monitoring at different levels (e.g., inverter, combiner, string, or module-level monitoring) is unclear since these depend on the quality of components, financial metrics, and climatic conditions. PVRPM was initially developed in Goldsim [@PVRPM:2011] [@PVRPM:2012] and then adapted in LK script [@osti_1761998] to improve the accuracy of energy yield simulations by inclusion of realistic energy losses and O&M costs [@PVRPM:EVAL]. However, these models lack many needed features such as efficient matrix operations, command-line interfaces, and external libraries to perform computations.
[…]
