### 2022-03-06 FSEC-Photovoltaics/pvrpm-lcoe#58 @brandons209
fixed in 75e52c7
