### 2022-03-06 FSEC-Photovoltaics/pvrpm-lcoe#60 @brandons209
fixed in 75e52c7 and thanks for the feedback!
