### 2022-03-06 FSEC-Photovoltaics/pvrpm-lcoe#58 @blthayer
@brandons209

> Will do, also I am a bit worried if your LCOE is in the thousands. Definitely should not be that high.

If I were to guess, it's related to the small system size that I used, combined with the default failure/monitoring configurations which are likely expensive for the small amount of energy output. Just speculating, though.
