### 2022-03-06 FSEC-Photovoltaics/pvrpm-lcoe#58 @brandons209
Yeah probably, just that even in systems I've seen its usually not that high, unless you had some super high amount of failures and availability was very low.
