### 2022-03-14 #4093 @jgostick
# Editor Task Prior to Acceptance
- [ ] Set article DOI with ``@whedon set <DOI here> as archive``
- [ ] Do a 'dry run' of acceptance with ``@whedon recommend accept``
