### 2022-03-14 #4093 @SarthakJariwala
> Ok, there is one checkbox left for @SarthakJariwala, but I think we can overlook based on @blthayer deep review. So, let's proceed with acceptance! I will do it this afternoon.

@jgostick I haven't been able to verify the functionality of the software primarily because it isn't running as expected on macOS. Otherwise, I am happy with everything else and agree you can overlook that given @blthayer review of the functionality.

@brandons209 I would recommend you add a note for your users in your docs (preferably in the installation section) that there are issues getting it to work on macOS. Based on our discussion in https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/issues/49 I agree it is likely a PySAM issue. Great work and good luck!
