### 2022-03-15 198fd74
add in messages about macOS support
