+++ README.md
@@ -20,4 +20,6 @@ Currently using SAM version 2021.12.02
 
 Please see the documentation for installation and usage guides.
 
+*Note:* Currently Py-PVRPM has issues running on macOS, pleas see #49 for updates.
+
 If you would like to contribute to PVRPM, have a question, or need support, please see the contributing guide [here](.github/CONTRIBUTING.md) for guides on submitting these issues.
