+++ .github/workflows/tests.yml
@@ -16,15 +16,15 @@ jobs:
           architecture: "x64"
       - run: pip install .[testing] && pytest
 
-  pytest-macos:
-    runs-on: macos-latest
-    steps:
-      - uses: actions/checkout@v2
-      - uses: actions/setup-python@v2
-        with:
-          python-version: "3.8.x"
-          architecture: "x64"
-      - run: pip install .[testing] && pytest
+#  pytest-macos:
+#    runs-on: macos-latest
+#    steps:
+#      - uses: actions/checkout@v2
+#      - uses: actions/setup-python@v2
+#        with:
+#          python-version: "3.8.x"
+#          architecture: "x64"
+#      - run: pip install .[testing] && pytest
 
   pytest-windows:
     runs-on: windows-latest
