### 2022-03-15 #4093 @brandons209
Version 1.7.4 is the final release version for the paper. All the authors are correct.

Here is the archive on Zenodo: [10.5281/zenodo.6355115](https://doi.org/10.5281/zenodo.6355115)

I have confirmed the title and everything else looks good for the paper!
