### 2022-03-16 #4093 @kthyng
@jgostick Is there a reason one checkbox is unchecked?

@brandons209 Please edit the metadata in your Zenodo archive so that the title and author list match your JOSS paper. Also it looks like the version listed here and in the Zenodo archive are not the same, so please rectify that and let me know what version you want here.
