### 2022-03-21 #4093 @danielskatz
@brandons209 - I'm suggesting some small language changes in https://github.com/FSEC-Photovoltaics/pvrpm-lcoe/pull/61 - please merge this or let me know what you disagree with.  Once we do this, and get the last checkbox issue resolve, we can move to final acceptance and publication
